pyproject.toml
[project]
name = "mcp-oauth2-demo"
version = "0.1.0"
description = "Hello world MCP server with OAuth2/OIDC authentication"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.2.4",
    "pyjwt[crypto]>=2.12.1",
]
